requirements.txt
pypdf2~=3.0.1
Pillow~=9.4.0